refatoração do arquivo de rotas a partir da criação de um serviço de calculo - #1
