+++ views/operacoes.ejs
@@ -26,7 +26,7 @@
                             <tr>
                                 <td><%= operacao.descricao %></td>
                                 <td>R$ <%= operacao.valor %></td>
-                                <td><%= operacao.sub %></td>
+                                <td>R$ <%= operacao.sub %></td>
                                 <td class="td-acoes">
                                     <a class="btn btn-link" href="/operacoes/edit/<%= operacao._id %>">Editar</a>
                                     <a class="btn btn-link" href="/operacoes/delete/<%= operacao._id %>">Excluir</a>
